requirements.txt
bandit==1.6.3
Django==2.2.17
psycopg2-binary==2.8.5
pylint==2.6.0
PyNaCl==1.4.0
pytz==2020.4
sqlparse==0.4.1

